backend/requirements.txt
adal
python-dotenv
flask_cors
flask
requests